Automatic .env file setup.
On passgen.py's first time being ran, it'll check the .env file for "FLAG=#src". If it is found, it'll automatically set up the file for you with a SALT and PASS (removing the FLAG variable as well). On passgen.py's 2nd run and any after, it'll check the .env file and find that "FLAG=#src" isn't there and skip to the normal functions and code like usual.

In the very odd chance that you would like new .env credentials, all you will need to do is add FLAG=#src back to the .env file and then run passgen.py again.