Update passgen.py
Updated;
If needed files aren't found if an option that that needs to use said files are picked, Passgen will download them from this repo.
Also some fixes for the exit() function are being tested out for compatibility with an upcoming idea I have that I want to do. "exit()" is now "sys.exit()".